// frontend/src/api/generated/models/SubscriptionPublic.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SubscriptionPublic = {
    entity_id: number;
    version: number;
    is_current: boolean;
    created_at: string;
    user_entity_id: number;
    target_type: string;
    target_entity_id: number;
    notification_types: Array<string>;
};